Name: OSProcess-dtl.117 Author: dtl Time: 9 February 2020, 3:48:58.591165 pm UUID: 8f1f5ec2-df00-4d80-89b9-463ac5e48685 Ancestors: OSProcess-dtl.116
When using the trunk UnixOSProcessPlugin (not the oscog one), all tests pass except for OSPipeTestCase>>testIsAtEndOfFile. This remaining failure is caused by stdio hacks in sqFileAtEnd() in the oscog FilePlugin support code. Add an assert description to the test to explain the source of the problem. Note - the various well-intentioned hacks to support stdio streams have no end of unintended side effects. Don't do that, it is just plain wrong.