Name: OSProcess-dtl.130 Author: dtl Time: 20 August 2022, 9:53:07.864275 am
OSProcess 4.7.1
Remove support for IOHandle (https://wiki.squeak.org/squeak/996).

The IOHandle sar is available on SqueakMap. It is out of date with
respect to current Squeak images, so interested experimenters should
try it first with a contempory Squeak image and an earlier version of
OSProcess (before 4.7.1).

Remove getSessionIdentiferFromSourcesFile which was used for support
of early VMs up to Squeak 3.8.

Name: OSProcess-eem.129
Author: eem
Time: 26 July 2022, 12:03:21.837177 pm

It is possible for grimReaper to be nil, in which case the added
guard avoids an MNU on startup.

Name: OSProcess-dtl.128
Author: dtl
Time: 6 July 2022, 11:58:57.621965 am

OSProcess 4.6.28 Fix a timing glitch in OSProcess class>>outputOf:
Use errorUpToEndOfFile to ensure full retrieval of error pipeline contents.